refactor(core-tests): 提取共享架构上下文测试基类
- 重构 TestArchitectureContext 与 TestArchitectureContextV3 的共享 IArchitectureContext 测试实现到 TestArchitectureContextBase
- 更新 gframework-pr-review 技能规则，要求对已验证成立的 nitpick 进行显式 triage
- 补充 analyzer-warning-reduction 的恢复点与验证记录
