+++ ai-plan/public/analyzer-warning-reduction/todos/analyzer-warning-reduction-tracking.md
@@ -6,13 +6,12 @@
 
 ## 当前恢复点
 
-- 恢复点编号：`ANALYZER-WARNING-REDUCTION-RP-090`
-- 当前阶段：`Phase 90`
+- 恢复点编号：`ANALYZER-WARNING-REDUCTION-RP-091`
+- 当前阶段：`Phase 91`
 - 当前焦点：
-  - `2026-04-28` 再次执行 `$gframework-pr-review`，确认 `PR #300` 最新 head 仍显示 `6` 条 CodeRabbit open threads，但本地复核后只有事件 API 回归覆盖仍然成立
-  - 已在 `TestArchitectureContextBehaviorTests.cs` 补齐 `UnRegisterEvent` 行为与 `SendEvent` / `RegisterEvent` / `UnRegisterEvent` 的空参数契约测试
-  - 已整理 `TestResourceLoader.cs` 的命名空间缩进，收口本轮顺手处理的局部格式噪音
-  - `dotnet format --verify-no-changes` 当前仍会暴露 `GFramework.Core.Tests` 项目里跨多个无关文件的既有 `FINALNEWLINE`、`CHARSET`、`WHITESPACE` 基线，本轮不扩展到整项目格式波次
+  - `2026-04-28` 继续收口 `PR #300` 的剩余 nitpick：已把 `TestArchitectureContext` / `TestArchitectureContextV3` 的重复 `IArchitectureContext` 测试实现抽到共享基类 `TestArchitectureContextBase`
+  - 已将 “本地验证后仍然成立的 nitpick 不得默认降级为可选项” 写入 `.agents/skills/gframework-pr-review/SKILL.md`
+  - `dotnet format --verify-no-changes` 的 `GFramework.Core.Tests` 既有 `FINALNEWLINE`、`CHARSET`、`WHITESPACE` 基线仍保持独立，不与本轮共享基类重构混提
 
 ## 当前活跃事实
 
@@ -20,12 +19,10 @@
 - 当前直接验证结果：
   - `dotnet build GFramework.Core.Tests/GFramework.Core.Tests.csproj -c Release`
     - 最新结果：成功；`0 Warning(s)`、`0 Error(s)`
-  - `dotnet test GFramework.Core.Tests/GFramework.Core.Tests.csproj -c Release --no-build --filter "FullyQualifiedName~TestArchitectureContextBehaviorTests"`
-    - 最新结果：成功；`9` 通过、`0` 失败
-  - `dotnet format GFramework.Core.Tests/GFramework.Core.Tests.csproj --verify-no-changes --no-restore`
-    - 最新结果：失败；暴露 `GFramework.Core.Tests` 项目中跨多个未触碰文件的既有 `FINALNEWLINE`、`CHARSET`、`WHITESPACE` 诊断，本轮新增写集未引入 `git diff --check` 异常
+  - `dotnet test GFramework.Core.Tests/GFramework.Core.Tests.csproj -c Release --no-build --filter "FullyQualifiedName~ArchitectureServicesTests|FullyQualifiedName~ContextAwareServiceExtensionsTests|FullyQualifiedName~TestArchitectureContextBehaviorTests|FullyQualifiedName~RegistryInitializationHookBaseTests|FullyQualifiedName~ArchitectureContextTests"`
+    - 最新结果：成功；`67` 通过、`0` 失败
 - 当前批次摘要：
-  - 当前工作树包含 `4` 个已修改文件，分别位于 `GFramework.Core.Tests` 与 `ai-plan/public/analyzer-warning-reduction`
+  - 当前工作树包含 `6` 个已修改文件与 `1` 个新增文件，分别位于 `GFramework.Core.Tests`、`.agents/skills/gframework-pr-review/` 与 `ai-plan/public/analyzer-warning-reduction`
   - 本轮没有触碰 `Mediator/*`、`YamlConfigSchemaValidator*` 或 `GFramework.Core.Tests` 的整项目格式基线波次
 
 ## 当前风险
@@ -60,6 +57,6 @@
 
 ## 下一步建议
 
-1. 提交本轮 `PR #300` review follow-up 与 `ai-plan` 同步。
-2. 若需要继续收口 PR 线程，可单独评估是否接受 `TestArchitectureContext` / `TestArchitectureContextV3` 的共享 helper nitpick。
+1. 提交本轮 `PR #300` nitpick follow-up、技能规则更新与 `ai-plan` 同步。
+2. 推送后重新执行 `$gframework-pr-review`，确认该 nitpick thread 是否随最新 head 自动收口。
 3. 若要清理 `dotnet format` 基线，另开 `GFramework.Core.Tests` 格式治理切片，不与当前 PR review 修复混提。
